requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
pydantic==2.9.0
pandas==2.2.2
python-dateutil==2.9.0